refactor(desktop): extract NODE_VERSION into shared module
Pull the bundled-Node version string and per-platform download
metadata out of scripts/download-nodejs.cjs and scripts/after-pack.cjs
into a new scripts/node-version.cjs module, and have both scripts
import from there.

Pure refactor — bundled Node stays at 22.22.2, SHA256 values
unchanged, archive filenames unchanged. Prevents silent drift where
the downloader fetches one version and after-pack looks for a
different one.
